+++ examples/hitechnic-accelerometer-SMUX-test1.c
@@ -44,9 +44,9 @@
 const tMUXSensor HTAC = msensor_S1_1;
 
 task main () {
-  int _x_axis = 0;
-  int _y_axis = 0;
-  int _z_axis = 0;
+  short _x_axis = 0;
+  short _y_axis = 0;
+  short _z_axis = 0;
 
   string _tmp;
