+++ examples/hitechnic-irseeker-v1-test1.c
@@ -28,8 +28,8 @@
 #include "hitechnic-irseeker-v1.h"
 
 task main () {
-  int _dir = 0;
-	int dcS1, dcS2, dcS3, dcS4, dcS5 = 0;
+  short _dir = 0;
+	short dcS1, dcS2, dcS3, dcS4, dcS5 = 0;
 
   displayCenteredTextLine(0, "HiTechnic");
   displayCenteredBigTextLine(1, "IR Seekr");
