+++ examples/hitechnic-irseeker-v2-SMUX-test1.c
@@ -42,23 +42,23 @@ const tMUXSensor HTIRS2 = msensor_S1_1;
 string sTextLines[8];
 
 // function prototypes
-void displayText(int nLineNumber, const string cChar, int nValueDC, int nValueAC);
+void displayText(short nLineNumber, const string cChar, short nValueDC, short nValueAC);
 void displayInstructions();
 
 // main task
 task main ()
 {
-  int _dirDC = 0;
-  int _dirAC = 0;
-	int dcS1, dcS2, dcS3, dcS4, dcS5 = 0;
-	int acS1, acS2, acS3, acS4, acS5 = 0;
+  short _dirDC = 0;
+  short _dirAC = 0;
+	short dcS1, dcS2, dcS3, dcS4, dcS5 = 0;
+	short acS1, acS2, acS3, acS4, acS5 = 0;
 	string tmpString;
 
   // show the user what to do
   displayInstructions();
 
   eraseDisplay();
-  for (int i = 0; i < 8; ++i)
+  for (short i = 0; i < 8; ++i)
     sTextLines[i] = "";
 
   // display the current DSP mode
@@ -117,7 +117,7 @@ void displayInstructions()
 }
 
 // Minimize LCD screen flicker by only updating LCD when data has changed
-void displayText(int nLineNumber, const string cChar, int nValueDC, int nValueAC)
+void displayText(short nLineNumber, const string cChar, short nValueDC, short nValueAC)
 {
   string sTemp;
