+++ examples/hitechnic-protoboard-SMUX-test2.c
@@ -38,7 +38,7 @@
 const tMUXSensor HTPB = msensor_S1_1;
 
 task main() {
-  int _chVal = 0;
+  short _chVal = 0;
 
   displayCenteredTextLine(0, "HiTechnic");
   displayCenteredBigTextLine(1, "Proto");
