+++ examples/hitechnic-protoboard-exp2.c
@@ -28,7 +28,7 @@
 
 task main() {
   // Local variables
-  int inputdata;
+  short inputdata;
   ubyte outputdata;
 
   // The data to be written: 0x3F = 111111 binary,
