+++ examples/lego-light-test2.c
@@ -25,8 +25,8 @@
 #include "lego-light.h"
 
 task main () {
-  int raw = 0;
-  int nrm = 0;
+  short raw = 0;
+  short nrm = 0;
   // Get control over the buttons
   nNxtButtonTask  = -2;
